Move alert generation for errors to actions/alert
